+++ controllers/warpMenuCreator.go
@@ -8,16 +8,16 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-// WarpMenuCreator used to create warp menu
-type WarpMenuCreator struct {
+// warpMenuCreator used to create warp menu
+type warpMenuCreator struct {
 	client    client.Client
 	registry  registry.Registry
 	namespace string
 }
 
 // NewWarpMenuCreator initialises a creator object to start the warp menu creation
-func NewWarpMenuCreator(client client.Client, registry registry.Registry, namespace string) WarpMenuCreator {
-	return WarpMenuCreator{
+func NewWarpMenuCreator(client client.Client, registry registry.Registry, namespace string) *warpMenuCreator {
+	return &warpMenuCreator{
 		client:    client,
 		registry:  registry,
 		namespace: namespace,
@@ -25,13 +25,13 @@ func NewWarpMenuCreator(client client.Client, registry registry.Registry, namesp
 }
 
 // Start starts the runnable.
-func (wmc WarpMenuCreator) Start(ctx context.Context) error {
+func (wmc warpMenuCreator) Start(ctx context.Context) error {
 	return wmc.CreateWarpMenu(ctx)
 }
 
 // CreateWarpMenu reads the warp configuration and starts watchers to refresh the menu.json configmap
 // in background.
-func (wmc WarpMenuCreator) CreateWarpMenu(ctx context.Context) error {
+func (wmc warpMenuCreator) CreateWarpMenu(ctx context.Context) error {
 	warpWatcher, err := warp.NewWatcher(ctx, wmc.client, wmc.registry, wmc.namespace)
 	if err != nil {
 		return fmt.Errorf("failed to create warp menu watcher: %w", err)